+++ tests/Exception/UnexpectedApiResultExceptionTest.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kyto\Alibaba\Tests\Exception;
+
+use Kyto\Alibaba\Exception\UnexpectedApiResultException;
+use PHPUnit\Framework\TestCase;
+
+class UnexpectedApiResultExceptionTest extends TestCase
+{
+    public function testConstruct(): void
+    {
+        $message = 'Message';
+        $code = 1;
+        $previous = new \RuntimeException('Previous');
+
+        $exception = new UnexpectedApiResultException($message, $code, $previous);
+
+        self::assertSame('Message', $exception->getMessage());
+        self::assertSame($code, $exception->getCode());
+        self::assertSame($previous, $exception->getPrevious());
+    }
+}
